Refactor and add null checks and tests
